requirements.txt
Flask==0.12.2
cf-deployment-tracker==1.0.4
cloudant==2.8.1
Pillow==3.3.3
boto3==1.6.18
tinys3==0.1.12
